restore authorization
