+++ lambdas/gateway/requirements.txt
@@ -5,3 +5,4 @@ mangum==0.17.0
 tiktoken==0.6.0
 requests==2.32.0
 numpy==1.26.4
+boto3==1.28.58
