--- lambdas/gateway_legacy/.gitignore
@@ -1,3 +1,0 @@
-bedrock-runtime/**
-test.py
-test/**
